requirements.txt
discord.py
PyGithub
databases[sqlite]
